backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
flask-session==0.8.0
gunicorn==23.0.0
pymongo==4.15.3
dnspython==2.8.0
python-dotenv==1.1.1
requests==2.32.3
Werkzeug==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
colorama==0.4.6
bcrypt==4.1.3

